pyproject.toml
[project]
name = "inventoria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.1",
    "bcrypt<4.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.1",
    "pandas-gbq>=0.29.2",
    "passlib[argon2,bcrypt]<2.0.0",
    "plotly>=5.18.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.47.0",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
